testing to use more than just headers
On windows only works if gdal.dll and its dependencies are copied next to libGeoRasters.dll.
This is undesired because it would mean that the gdextension also needs to ship all those dependencies.

Might after all need to switch to manually compile gdal to have it fully statically compiled into libGeoRasters.dll.